Fix refresh token request

+++ src/utils/axios.js
@@ -27,20 +27,36 @@ axiosApiInstance.interceptors.response.use(
     return response;
   },
   function (error) {
-    const refreshToken = localStorage.getItem("refreshToken");
-    axiosApiInstance
-      .post("auth/refresh", { refreshToken })
-      .then((res) => {
-        localStorage.setItem("token", res.data.data.token);
-        localStorage.setItem("refreshToken", res.data.data.refreshToken);
-        window.location.reload();
-      })
-      .catch(() => {
-        localStorage.clear();
-        window.location.href = "/basic/login";
-      });
     // Any status codes that falls outside the range of 2xx cause this function to trigger
     // Do something with response error
+    if (error.response.status === 403) {
+      if (error.response.data.msg !== "jwt expired") {
+        localStorage.clear();
+        window.location.href = "/basic/login";
+      } else {
+        const refreshToken = localStorage.getItem("refreshToken");
+        axiosApiInstance
+          .post("auth/refresh", { refreshToken })
+          .then((res) => {
+            // res = {
+            //   data: {
+            //     data: {
+            //       id: ...
+            //       token: ...
+            //       refreshToken: ...
+            //     }
+            //   }
+            // }
+            localStorage.setItem("token", res.data.data.token);
+            localStorage.setItem("refreshToken", res.data.data.refreshToken);
+            window.location.reload();
+          })
+          .catch(() => {
+            localStorage.clear();
+            window.location.href = "/basic/login";
+          });
+      }
+    }
     return Promise.reject(error);
   }
 );
